requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
idna==2.6
oauthlib==2.0.2
psycopg2==2.7.1
python-dotenv==0.6.5
pytz==2017.2
requests==2.18.4
urllib3==1.22
